refactor: Collectionview를 custom class로 분리 #7
- ProductCollectionView class 생성
- mainViewController내 collectionview 로직 분리
- UIsrcollview extension 코드 분리
- mainViewController내 불필요한 코드 제거
